+++ packages/cli/src/cli.ts
@@ -11,16 +11,25 @@ import { getVersion } from './utils/get-version';
 import { getDockerComposeVersion } from './utils/get-docker-compose-version';
 import { formatLintResults, lint } from './linter/lint';
 import path from 'path';
+import {Context} from "./types";
 
-const context = createContext('local-api-gateway.yml');
 const minDockerComposeVersion = '1.25.5';
+let loadedContext: Context|null = null;
+
+const loadContext = () => {
+    if (loadedContext === null) {
+        loadedContext = createContext('local-api-gateway.yml');
+    }
+
+    return loadedContext;
+};
 
 const dockerComposePassthrough = async (args: string[]) => {
     try {
         const command = `docker-compose ${args.join(' ')}`;
 
         console.log('Executing command:', chalk.black.bgWhite(command));
-        await execa(command, { shell: true, stdio: 'inherit', cwd: context.directories.build });
+        await execa(command, { shell: true, stdio: 'inherit', cwd: loadContext().directories.build });
     } catch (error) {
         console.log(error.message);
     }
@@ -60,6 +69,7 @@ program
         const args = [...program.args];
         const dockerComposeVersion = await getDockerComposeVersion();
         const noBuildConfigIndex = args.indexOf('--no-build-config');
+        const context = loadContext();
 
         console.log('Local API gateway version ', chalk.black.bgWhite(getVersion()));
         console.log('Docker compose version ', chalk.black.bgWhite(dockerComposeVersion || 'unknown'));