+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 0.3.1 - 2021-05-09
+### Fixed
+ - CLI: Lint command no longer requires checking out the integration sources to run.
+
 ## 0.3.0 - 2021-05-09
 ### Added
  - CLI: Added `lint` command for validating structure and some other aspects of `local-api-gateway.yml`.